// src/app/api/models/mesa-electoral.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { TipoMesa } from '../models/tipo-mesa';
export interface MesaElectoral {
  activa?: boolean;
  codigoMesa?: string | null;
  descripcion?: string | null;
  eleccionId?: string;
  idMesaElectoral?: string;
  tipoMesa?: TipoMesa;
  zonaId?: string;
}